requirements.txt
scipy==1.4.*
numpy==1.17.*
pandas==1.0.*
PyYAML==5.3.*
tensorboard==2.0.2
torch==1.3.1
tqdm==4.43.*
transformers==2.3.0
fairseq==0.9.*
pytorch-lightning==0.7.1
test-tube==0.7.5
pytorch-nlp==0.5.0
scikit-learn==0.22.2.post1
